refactor(artifacts-importer): split guide and import components
